Confidence Intervals for Conditional Effects and Evaluation Measures

Wald-style confidence intervals for data.tables with additional columns for estimates of the conditional average treatment effect (CATE) and related segmentation evaluation measures.

ci_wald(estimation_results, segment_by = NULL, param_type = c("cate",
  "effect_measure"), outcome_type = c("continuous", "binary"),
  coverage = getOption("sherlock.ci_covers"))

Arguments

estimation_results
A data.table containing the input data, augmented to include cross-validated nuisance parameter estimates, an estimate of the CATE, and a treatment rule assigned based on the estimated CATE via assign_rule. This input object should be generated by successive calls to set_est_data and est_cate, or through a wrapper function that composes these function calls automatically. For cases in which assign_rule has been called prior, any additional information is preserved.

segment_by
A character vector specifying the column names in data_obs that correspond to the covariates over which segmentation should be performed. This should be a strict subset of baseline.

param_type
A character string (of length one) indicating the type of estimates for which a Wald-style confidence interval is to be computed. The choice of "cate" indicates that this routine is being invoked thru summarize_segments while "effect_measure" indicates that the input was generated by est_effect.

outcome_type
A character string (of length one) indicating whether the outcome variable is binary or continuous-valued. For cases in which the outcome is binary, the confidence interval for the parameter estimate is constructed on the logit scale and then back-transformed, in order to respect the fact that the estimate must lie in the unit interval.

coverage
A numeric indicating the nominal rate at which the Wald-style confidence interval is intended to cover the target parameter.

Value

A data.table matching exactly the input provided in estimation_results, augmented to include the lower and upper bounds from the Wald-style confidence intervals constructed.