Draw Cartesian Small Multiples Plot

Draws a cartesian small multiples plot

Usage

draw_cartesian_small_multiples(
  data,
  x_coord,
  y_coord,
  grouping_var,
  faceting_var_1,
  faceting_var_2,
  interactive = FALSE,
  size = 2,
  alpha = 0.4,
  analysis_desc_label = NULL,
  x_axis_label = NULL,
  y_axis_label = NULL,
  n_breaks_x_axis = 10,
  n_breaks_y_axis = 10,
  accuracy = 0.001,
  show_axis_values = TRUE
)

Arguments

data

Input dataset to be plotted (required)

x_coord

Column for X coordinate values (required)

y_coord

Column for Y coordinate values (required)

grouping_var

Grouping variable. Each group is displayed in a different color. (optional)

faceting_var_1

Set first faceting variable (optional)

faceting_var_2

Set second faceting variable (optional)

interactive

Set plot interactivity. By default, it is set to FALSE (optional)

size

Set point size. By default, it is set to 2 (optional)

alpha

Set transparency. By default, it is set to 0.4 (optional)

analysis_desc_label

Label (subtitle) for analysis description. By default, it is set to NULL (optional)

x_axis_label

Label for x axis. By default, it is set to display x axis column name (optional)

y_axis_label

Label for y axis. By default, it is set to display y axis column name (optional)

n_breaks_x_axis

Set number of breaks on X axis. By default, it is set to 10 (optional)

n_breaks_y_axis

Set number of breaks on Y axis. By default, it is set to 10 (optional)

accuracy

Set number of decimal places to be displayed on X and Y axes. Examples: 0.1 - one decimal place, 0.01 - two decimal places, 0.001 - three decimal places etc. By default, it is set to 0.001 (optional)

show_axis_values

Logical. if FALSE, default, axis values are not shown (optional)

Value

A 'ggplot' or 'plotly' object